Changes to reconciliation in transactions in guide
Minor reorganisation of reconcilation section to put the reconcile status info in a list form. Add info re dates not necessarily matching and sequence of transactions being different in reconcile window panes and on statement. Add additional section on transactions which are in GnuCash record but not on statement. Add to tips re toolbar buttons and using mouse to check and uncheck transactions.  Add an additional section on general reconciliation problems and tips for recovering deleted transactions etc.